+++ LibOS/shim/src/Makefile
@@ -8,7 +8,8 @@ CFLAGS += -fPIC -Winline -Wwrite-strings \
 	  $(cc-option, -Wnull-dereference) \
 	  -fno-stack-protector -fno-builtin -Wno-inline \
 	  -I../include -I../../../Pal/include/lib -I../../../Pal/include/pal \
-	  -I../../../Pal/include/elf -I../../../Pal/include/lib/$(ARCH)
+	  -I../../../Pal/include/elf -I../../../Pal/include/lib/$(ARCH) \
+	  -I../../../Pal/include/arch/$(ARCH)
 
 CFLAGS += -Wextra
