+++ LibOS/shim/test/inline/Makefile
@@ -18,6 +18,8 @@ include ../../../../Scripts/Makefile.configs
 include ../../../../Scripts/Makefile.manifest
 include ../../../../Scripts/Makefile.Test
 
+CFLAGS += -I $(PALDIR)/../include/arch/$(ARCH)
+
 .PHONY: crt_init-recurse
 crt_init-recurse:
 	$(MAKE) -C ../crt_init all
