+++ Pal/lib/Makefile
@@ -2,7 +2,8 @@ include ../../Scripts/Makefile.configs
 include ../../Scripts/Makefile.rules
 include ../src/host/$(PAL_HOST)/Makefile.am
 
-CFLAGS += -I../include/lib -I../include -I../include/pal -I../include/host/$(PAL_HOST) \
+CFLAGS += -I../include/lib -I../include -I../include/pal -I../include/arch/$(ARCH) \
+          -I../include/host/$(PAL_HOST) \
           -I../src/host/$(PAL_HOST) -Icrypto/mbedtls/include -Icrypto/mbedtls/crypto/include
 
 CRYPTO_PROVIDER ?= mbedtls
